Move ServiceInfo into _types.py.
