+++ packages/ni.measurementlink.discovery.v1.client/src/ni/measurementlink/discovery/v1/client/__init__.py
@@ -1,9 +1,9 @@
 """Public API for accessing the NI Discovery Service."""
 
 from ni.measurementlink.discovery.v1.client._client import DiscoveryClient
-from ni.measurementlink.discovery.v1.client._info import ServiceInfo
 from ni.measurementlink.discovery.v1.client._types import (
     ComputeNodeDescriptor,
+    ServiceInfo,
     ServiceLocation,
 )
